code tweak for iOS
